+++ src/resources/views/order/show/_details.blade.php
@@ -1,12 +1,16 @@
 <div class="card mb-3">
     <div class="card-header">
-        {{ __('Additonal Details') }}
+        {{ __('Additional Details') }}
     </div>
 
     <div class="card-body">
         <h6>{{ __('Customer Notes') }}</h6>
         <div class="font-italic">
-            {!! nl2br($order->notes) !!}
+            @empty($order->notes)
+                -
+            @else
+                {!! nl2br($order->notes) !!}
+            @endempty
         </div>
     </div>
 </div>